+++ RELEASE-NOTES.md
@@ -7,6 +7,7 @@
       is available for auto-loading
 * #46 Support classpath:/ URLs in configuration loader.
 * #47 Configuration type for abstract resource
+* #54 @BindingAnnotation's in Bootique should be allowed on methods
 
 ## 0.14: